+++ src/app.rs
@@ -1,11 +1,11 @@
 use crate::{
-    action::Action,
     build::{self, BodyComponents},
     components::Component,
     config::Config,
     focus::FocusManager,
-    ros_ext::RocksDBSwitch,
     shared::{Shared, shared},
+    states::Action,
+    store::rocks_db_switch::RocksDBSwitch,
     tui::{Event, Tui},
 };
 use color_eyre::Result;
