+++ src/components/list.rs
@@ -1,9 +1,9 @@
 use crate::{
-    action::Action,
+    states::Action,
     components::Component,
     focus::{FocusState, FocusableComponent},
     shared::GetterOpt,
-    to_list_item::ToListItem,
+    ui::to_list_item::ToListItem,
     window::WindowState,
 };
 use color_eyre::Result;
