+++ src/components/static/search_types.rs
@@ -1,4 +1,4 @@
-use crate::to_list_item::ToListItem;
+use crate::ui::to_list_item::ToListItem;
 use ratatui::widgets::ListItem;
 use serde::Serialize;
 use strum::EnumIter;
