--- src/nav.rs
@@ -1,7 +1,0 @@
-use strum::{Display, EnumIter};
-
-#[derive(Clone, Debug, EnumIter, Display, PartialEq, Eq)]
-pub enum NavMode {
-    Browse,
-    Search,
-}
