+++ src/ui/to_rich/proposal.rs
@@ -1,6 +1,6 @@
 use crate::{
-    to_list_item::ProposalItem,
-    to_rich::{
+    ui::to_list_item::ProposalItem,
+    ui::to_rich::{
         RationalNumberDisplay, ToRichText, labeled_default, labeled_default_opt,
         labeled_default_opt_single, labeled_default_single,
     },
